+++ src/app/modules/inference/components/inferences-list/inferences-list.component.spec.ts
@@ -1,0 +1,59 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { InferencesListComponent } from './inferences-list.component';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { SharedModule } from '@app/shared/shared.module';
+import { MediaMatcher } from '@angular/cdk/layout';
+import { TranslateModule } from '@ngx-translate/core';
+import { AppConfigService } from '@app/core/services/app-config/app-config.service';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+
+const mockedConfigService: any = {};
+
+const mockedMediaQueryList: MediaQueryList = {
+    matches: true,
+    media: 'test',
+    onchange: jest.fn(),
+    addListener: jest.fn(),
+    removeListener: jest.fn(),
+    addEventListener: jest.fn(),
+    dispatchEvent: jest.fn(),
+    removeEventListener: jest.fn(),
+};
+
+const mockedMediaMatcher: any = {
+    matchMedia: jest.fn().mockReturnValue(mockedMediaQueryList),
+};
+
+describe('ServicesListComponent', () => {
+    let component: InferencesListComponent;
+    let fixture: ComponentFixture<InferencesListComponent>;
+
+    beforeEach(async () => {
+        await TestBed.configureTestingModule({
+            declarations: [InferencesListComponent],
+            imports: [
+                SharedModule,
+                NoopAnimationsModule,
+                HttpClientTestingModule,
+                TranslateModule.forRoot(),
+            ],
+            providers: [
+                { provide: AppConfigService, useValue: mockedConfigService },
+
+                { provide: MatDialogRef, useValue: {} },
+                { provide: MAT_DIALOG_DATA, useValue: {} },
+                { provide: MediaMatcher, useValue: mockedMediaMatcher },
+            ],
+        }).compileComponents();
+
+        fixture = TestBed.createComponent(InferencesListComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
+
+    it('should create', () => {
+        expect(component).toBeTruthy();
+    });
+});