+++ src/app/shared/components/popup/popup/popup.component.ts
@@ -1,0 +1,15 @@
+import { Component, Inject } from '@angular/core';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+
+@Component({
+    selector: 'app-popup',
+    templateUrl: './popup.component.html',
+    styleUrls: ['./popup.component.scss'],
+})
+export class PopupComponent {
+    constructor(
+        public dialogRef: MatDialogRef<PopupComponent>,
+        @Inject(MAT_DIALOG_DATA)
+        public data: { message: string }
+    ) {}
+}