+++ src/app/shared/interfaces/platform-status.interface.ts
@@ -1,0 +1,16 @@
+export interface PlatformStatus {
+    id: number;
+    title: string;
+    vo?: string;
+    body?: string;
+    start?: Date;
+    end?: Date;
+}
+
+export interface StatusNotification {
+    title: string;
+    vo?: string;
+    summary?: string;
+    start?: Date;
+    end?: Date;
+}